PAE-1411: consolidate prn-summary tests via describe.each
