+++ src/server/reports/exporter/prn-summary-controller.test.js
@@ -1,40 +1,28 @@
 import { config } from '#config/config.js'
 import { statusCodes } from '#server/common/constants/status-codes.js'
 import { fetchRegistrationAndAccreditation } from '#server/common/helpers/organisations/fetch-registration-and-accreditation.js'
-import { getCsrfToken } from '#server/common/test-helpers/csrf-helper.js'
 import { fetchReportDetail } from '#server/reports/helpers/fetch-report-detail.js'
 import { it } from '#vite/fixtures/server.js'
-import { getByRole, getByText } from '@testing-library/dom'
-import { JSDOM } from 'jsdom'
 import { afterAll, beforeAll, beforeEach, describe, expect, vi } from 'vitest'
 
 vi.mock(
   import('#server/common/helpers/organisations/fetch-registration-and-accreditation.js')
 )
 vi.mock(import('#server/reports/helpers/fetch-report-detail.js'))
-vi.mock(import('../helpers/update-report.js'))
-
-const { updateReport } = await import('../helpers/update-report.js')
 
 const mockCredentials = {
-  profile: {
-    id: 'user-123',
-    email: 'test@example.com'
-  },
+  profile: { id: 'user-123', email: 'test@example.com' },
   idToken: 'mock-id-token'
 }
 
-const mockAuth = {
-  strategy: 'session',
-  credentials: mockCredentials
-}
+const mockAuth = { strategy: 'session', credentials: mockCredentials }
 
-const accreditedExporter = {
+const accreditedReprocessor = {
   organisationData: { id: 'org-123' },
   registration: {
     id: 'reg-001',
     material: 'plastic',
-    wasteProcessingType: 'exporter',
+    wasteProcessingType: 'reprocessor',
     registrationNumber: 'REG001234'
   },
   accreditation: {
@@ -43,21 +31,8 @@ const accreditedExporter = {
   }
 }
 
-const registeredOnlyExporter = {
-  ...accreditedExporter,
-  accreditation: undefined
-}
-
-const accreditedReprocessor = {
-  ...accreditedExporter,
-  registration: {
-    ...accreditedExporter.registration,
-    wasteProcessingType: 'reprocessor'
-  }
-}
-
 const reportDetail = {
-  operatorCategory: 'EXPORTER_ACCREDITED',
+  operatorCategory: 'REPROCESSOR_ACCREDITED',
   cadence: 'monthly',
   year: 2026,
   period: 1,
@@ -70,481 +45,49 @@ const reportDetail = {
   status: { currentStatus: 'in_progress' },
   supportingInformation: null,
   recyclingActivity: {
-    totalTonnageReceived: 80.25,
+    totalTonnageReceived: 200,
     suppliers: [],
     tonnageRecycled: null,
     tonnageNotRecycled: null
   },
-  wasteSent: {
-    tonnageSentToReprocessor: 5,
-    tonnageSentToExporter: 3,
-    tonnageSentToAnotherSite: 2,
-    finalDestinations: []
-  },
   prn: {
     issuedTonnage: 91,
     totalRevenue: null,
+    freeTonnage: null,
     averagePricePerTonne: null
   }
 }
 
-const reportDetailWithRevenue = {
-  ...reportDetail,
-  prn: {
-    ...reportDetail.prn,
-    totalRevenue: 1576.12
-  }
-}
+const baseUrl =
+  '/organisations/org-123/registrations/reg-001/reports/2026/monthly/1/prn-summary'
 
-const organisationId = 'org-123'
-const registrationId = 'reg-001'
-const baseUrl = `/organisations/${organisationId}/registrations/${registrationId}/reports/2026/monthly/1/prn-summary`
+describe('#prnSummaryDispatcher cross-registration routing', () => {
+  beforeAll(() => {
+    config.set('featureFlags.reports', true)
+  })
 
-describe('#prnSummaryController', () => {
   beforeEach(() => {
     vi.clearAllMocks()
   })
 
-  describe('when feature flag is enabled', () => {
-    beforeAll(() => {
-      config.set('featureFlags.reports', true)
-    })
-
-    afterAll(() => {
-      config.reset('featureFlags.reports')
-    })
-
-    describe('GET', () => {
-      beforeEach(() => {
-        vi.mocked(fetchRegistrationAndAccreditation).mockResolvedValue(
-          accreditedExporter
-        )
-        vi.mocked(fetchReportDetail).mockResolvedValue(reportDetail)
-      })
-
-      it('should return 200 for accredited exporter with monthly cadence', async ({
-        server
-      }) => {
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        expect(statusCode).toBe(statusCodes.ok)
-      })
-
-      it('should display the heading', async ({ server }) => {
-        const { result } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        const { body } = new JSDOM(result).window.document
-
-        expect(
-          getByRole(body, 'heading', {
-            level: 1,
-            name: /PERNs issued for plastic packaging waste in January$/
-          })
-        ).toBeDefined()
-      })
-
-      it('should display tonnage issued', async ({ server }) => {
-        const { result } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        const { body } = new JSDOM(result).window.document
-        const inset = body.querySelector('.govuk-inset-text')
-
-        expect(inset).not.toBeNull()
-        expect(getByText(inset, /Total tonnage of PERNs issued/)).toBeDefined()
-        expect(getByText(inset, '91')).toBeDefined()
-      })
-
-      it('should pre-fill revenue if previously saved', async ({ server }) => {
-        vi.mocked(fetchReportDetail).mockResolvedValue(reportDetailWithRevenue)
-
-        const { result } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        expect(result).toContain('1576.12')
-      })
-
-      it.for([
-        { revenue: 1576.12, expected: '1576.12' },
-        { revenue: 1576.1, expected: '1576.10' },
-        { revenue: 1576, expected: '1576' }
-      ])(
-        'should pre-fill revenue $revenue as $expected',
-        async ({ revenue, expected }, { server }) => {
-          vi.mocked(fetchReportDetail).mockResolvedValue({
-            ...reportDetail,
-            prn: { ...reportDetail.prn, totalRevenue: revenue }
-          })
-
-          const { result } = await server.inject({
-            method: 'GET',
-            url: baseUrl,
-            auth: mockAuth
-          })
-
-          const { body } = new JSDOM(result).window.document
-          const input = getByRole(body, 'textbox', {
-            name: /Enter total revenue of PERNs, excluding VAT/
-          })
-
-          expect(input.value).toBe(expected)
-        }
-      )
-
-      it('should return 404 for non-accredited exporter', async ({
-        server
-      }) => {
-        vi.mocked(fetchRegistrationAndAccreditation).mockResolvedValue(
-          registeredOnlyExporter
-        )
-
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        expect(statusCode).toBe(statusCodes.notFound)
-      })
-
-      it('should dispatch to reprocessor controller for accredited reprocessor', async ({
-        server
-      }) => {
-        vi.mocked(fetchRegistrationAndAccreditation).mockResolvedValue(
-          accreditedReprocessor
-        )
-
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        expect(statusCode).toBe(statusCodes.ok)
-      })
-
-      it('should return 404 for quarterly cadence', async ({ server }) => {
-        const quarterlyUrl = `/organisations/${organisationId}/registrations/${registrationId}/reports/2026/quarterly/1/prn-summary`
-
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: quarterlyUrl,
-          auth: mockAuth
-        })
-
-        expect(statusCode).toBe(statusCodes.notFound)
-      })
-
-      it('should return 500 when prn data is missing', async ({ server }) => {
-        vi.mocked(fetchReportDetail).mockResolvedValue({
-          ...reportDetail,
-          prn: undefined
-        })
-
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        expect(statusCode).toBe(statusCodes.internalServerError)
-      })
-
-      it('should return 404 when report is not in progress', async ({
-        server
-      }) => {
-        vi.mocked(fetchReportDetail).mockResolvedValue({
-          ...reportDetail,
-          status: { currentStatus: 'ready_to_submit' }
-        })
-
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: baseUrl,
-          auth: mockAuth
-        })
-
-        expect(statusCode).toBe(statusCodes.notFound)
-      })
-    })
-
-    describe('POST', () => {
-      beforeEach(() => {
-        vi.mocked(fetchRegistrationAndAccreditation).mockResolvedValue(
-          accreditedExporter
-        )
-        vi.mocked(fetchReportDetail).mockResolvedValue(reportDetail)
-        vi.mocked(updateReport).mockResolvedValue(undefined)
-      })
-
-      describe('csrf protection', () => {
-        it('should reject POST without CSRF token', async ({ server }) => {
-          const { statusCode } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            payload: {}
-          })
-
-          expect(statusCode).toBe(statusCodes.forbidden)
-        })
-      })
-
-      describe('when continue is clicked with valid revenue', () => {
-        it('should redirect to free-perns', async ({ server }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          const { statusCode, headers } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: 1576.12, action: 'continue' }
-          })
-
-          expect(statusCode).toBe(statusCodes.found)
-          expect(headers.location).toBe(
-            `/organisations/${organisationId}/registrations/${registrationId}/reports/2026/monthly/1/free-perns`
-          )
-
-          expect(server.loggerMocks.info).toHaveBeenCalledWith({
-            message: 'prn-summary dispatch POST',
-            event: {
-              action: 'prn_summary_dispatch_post',
-              reason: 'reprocessor=false'
-            }
-          })
-        })
-
-        it('should call updateReport with prnRevenue', async ({ server }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: 1576.12, action: 'continue' }
-          })
-
-          expect(updateReport).toHaveBeenCalledWith(
-            organisationId,
-            registrationId,
-            2026,
-            'monthly',
-            1,
-            { prnRevenue: 1576.12 },
-            'mock-id-token'
-          )
-        })
-      })
-
-      describe('when save is clicked', () => {
-        it('should redirect to reports list', async ({ server }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          const { statusCode, headers } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: 1576.12, action: 'save' }
-          })
-
-          expect(statusCode).toBe(statusCodes.found)
-          expect(headers.location).toBe(
-            `/organisations/${organisationId}/registrations/${registrationId}/reports`
-          )
-        })
-
-        it('should redirect to reports list when revenue is empty', async ({
-          server
-        }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          const { statusCode, headers } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: '', action: 'save' }
-          })
-
-          expect(statusCode).toBe(statusCodes.found)
-          expect(headers.location).toBe(
-            `/organisations/${organisationId}/registrations/${registrationId}/reports`
-          )
-        })
-
-        it('should not call updateReport when revenue is empty', async ({
-          server
-        }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: '', action: 'save' }
-          })
-
-          expect(updateReport).not.toHaveBeenCalled()
-        })
-
-        it('should show error when revenue is non-numeric', async ({
-          server
-        }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          const { result } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: 'abc', action: 'save' }
-          })
-
-          const { body } = new JSDOM(result).window.document
-          const alert = getByRole(body, 'alert')
-
-          expect(
-            getByText(
-              alert,
-              /Enter the total revenue in digits, using a decimal point if needed/
-            )
-          ).toBeDefined()
-        })
-      })
-
-      describe('validation errors', () => {
-        it('should show error when revenue is non-numeric', async ({
-          server
-        }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          const { result } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: 'abc', action: 'continue' }
-          })
-
-          const { body } = new JSDOM(result).window.document
-          const alert = getByRole(body, 'alert')
-
-          expect(
-            getByText(
-              alert,
-              /Enter the total revenue in digits, using a decimal point if needed/
-            )
-          ).toBeDefined()
-
-          expect(server.loggerMocks.error).toHaveBeenCalledWith({
-            message: 'prn-summary dispatch validation failed',
-            event: {
-              action: 'prn_summary_dispatch_validation_failed',
-              reason: expect.any(String)
-            }
-          })
-        })
-
-        it('should show error when revenue has more than 2 decimal places', async ({
-          server
-        }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          const { result } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: 1576.123, action: 'continue' }
-          })
-
-          const { body } = new JSDOM(result).window.document
-          const alert = getByRole(body, 'alert')
-
-          expect(
-            getByText(
-              alert,
-              /Total revenue should only have 2 digits after the decimal point/
-            )
-          ).toBeDefined()
-        })
-
-        it('should show error when revenue is empty', async ({ server }) => {
-          const { cookie, crumb } = await getCsrfToken(server, baseUrl, {
-            auth: mockAuth
-          })
-
-          const { result } = await server.inject({
-            method: 'POST',
-            url: baseUrl,
-            auth: mockAuth,
-            headers: { cookie },
-            payload: { crumb, prnRevenue: '', action: 'continue' }
-          })
-
-          const { body } = new JSDOM(result).window.document
-          const alert = getByRole(body, 'alert')
-
-          expect(
-            getByText(alert, /Enter the total revenue of PERNs issued/)
-          ).toBeDefined()
-        })
-      })
-    })
+  afterAll(() => {
+    config.reset('featureFlags.reports')
   })
 
-  describe('when feature flag is disabled', () => {
-    beforeAll(() => {
-      config.set('featureFlags.reports', false)
+  it('should dispatch to reprocessor controller for accredited reprocessor', async ({
+    server
+  }) => {
+    vi.mocked(fetchRegistrationAndAccreditation).mockResolvedValue(
+      accreditedReprocessor
+    )
+    vi.mocked(fetchReportDetail).mockResolvedValue(reportDetail)
+
+    const { statusCode } = await server.inject({
+      method: 'GET',
+      url: baseUrl,
+      auth: mockAuth
     })
 
-    afterAll(() => {
-      config.reset('featureFlags.reports')
-    })
-
-    it('should return 404', async ({ server }) => {
-      const { statusCode } = await server.inject({
-        method: 'GET',
-        url: baseUrl,
-        auth: mockAuth
-      })
-
-      expect(statusCode).toBe(statusCodes.notFound)
-    })
+    expect(statusCode).toBe(statusCodes.ok)
   })
 })
